# services/copilot-adapter/openapi.yaml
openapi: 3.0.3
info:
  title: Copilot Adapter API
  version: 1.1.0
  description: |
    HTTP contract for the Copilot operator workbench: a REST surface over
    GitHub Copilot CLI sessions, bounded git worktrees, terminals, durable chat
    schedules, and the subagents reported by a live session. Every 4xx and 5xx
    response carries the shared Error schema. Identifiers are UUIDs unless a
    schema says an upstream identifier is retained, and every timestamp is an
    RFC 3339 instant in UTC.
servers:
  - url: http://127.0.0.1:8090
    description: The adapter running as a bare binary on loopback.
security:
  - {}
  - bearerAuth: []
...
paths:
...
  /v1/worktrees/{worktreeId}:
    parameters:
      - $ref: '#/components/parameters/WorktreeId'
    get:
      operationId: getWorktree
      summary: Read current worktree metadata
      description: |
        The branch, HEAD, cleanliness and presence are inspected from git for
        this response; they are not stale creation-time labels.
      tags:
        - worktrees
      responses:
        '200':
          description: The worktree and its current git metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worktree'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/worktrees/{worktreeId}/terminals:
    parameters:
      - $ref: '#/components/parameters/WorktreeId'
    get:
      operationId: listTerminals
      summary: List terminal sessions for a worktree
      tags:
        - terminals
      responses:
        '200':
          description: Live and recently exited terminals, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createTerminal
      summary: Start a bounded terminal in a worktree
      description: |
        Starts the configured shell with its working directory fixed to the
        worktree. The server owns the child process, output bound and cleanup;
        clients cannot choose an executable or a different directory.
      tags:
        - terminals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTerminalRequest'
      responses:
        '201':
          description: The terminal was started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/worktrees/{worktreeId}/terminals/{terminalId}:
    parameters:
      - $ref: '#/components/parameters/WorktreeId'
      - $ref: '#/components/parameters/TerminalId'
    get:
      operationId: getTerminal
      summary: Read terminal lifecycle state
      tags:
        - terminals
      responses:
        '200':
          description: The terminal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      operationId: resizeTerminal
      summary: Resize a running terminal
      tags:
        - terminals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResizeTerminalRequest'
      responses:
        '200':
          description: The resized terminal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: stopTerminal
      summary: Stop a terminal and release its process
      tags:
        - terminals
      responses:
        '200':
          description: The terminal after bounded shutdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/worktrees/{worktreeId}/terminals/{terminalId}/input:
    parameters:
      - $ref: '#/components/parameters/WorktreeId'
      - $ref: '#/components/parameters/TerminalId'
    post:
      operationId: writeTerminalInput
      summary: Write input to a running terminal
      tags:
        - terminals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalInputRequest'
      responses:
        '202':
          description: The bounded input chunk was accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Terminal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/worktrees/{worktreeId}/terminals/{terminalId}/events:
    parameters:
      - $ref: '#/components/parameters/WorktreeId'
      - $ref: '#/components/parameters/TerminalId'
    get:
      operationId: streamTerminalEvents
      summary: Stream bounded terminal output and exit state
      description: |
        A text/event-stream of TerminalEvent envelopes. Last-Event-ID resumes
        from the terminal's bounded replay buffer. A cursor older than retained
        output receives the oldest retained event with replayTruncated true.
      tags:
        - terminals
      parameters:
        - $ref: '#/components/parameters/LastEventId'
      responses:
        '200':
          description: Terminal event stream; each data frame is one TerminalEvent.
          content:
            text/event-stream:
              schema:
                type: string
                description: A newline-delimited SSE frame sequence.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    WorktreeId:
      name: worktreeId
      in: path
      required: true
      description: Adapter-owned worktree identifier.
      schema:
        type: string
        format: uuid
    TerminalId:
      name: terminalId
      in: path
      required: true
      description: Terminal identifier.
      schema:
        type: string
        format: uuid
...
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      description: |
        The seq of the last event the client already holds. The stream resumes
        immediately after it.
      schema:
        type: integer
        format: int64
        minimum: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Optional. The adapter binds loopback and requires no credential there;
        a deployment that fronts it with a reverse proxy may require a bearer
        token, which is why every operation accepts either.
  schemas:
...
    WorktreeState:
      type: string
      description: |
        active is present on disk; missing means its recorded path disappeared;
        released is an adapter-managed worktree that was intentionally removed.
      enum:
        - active
        - missing
        - released
    Worktree:
      type: object
      additionalProperties: false
      required:
        - id
        - repositoryId
        - repositoryRoot
        - path
        - branch
        - headSha
        - baseRef
        - managed
        - clean
        - state
        - sessionCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        repositoryId:
          type: string
          minLength: 1
          maxLength: 128
          readOnly: true
        repositoryRoot:
          type: string
          minLength: 1
          maxLength: 4096
          readOnly: true
        path:
          type: string
          minLength: 1
          maxLength: 4096
          readOnly: true
        branch:
          type: string
          maxLength: 1024
          readOnly: true
          description: Empty for a detached HEAD or a missing or released worktree.
        headSha:
          type: string
          pattern: '^$|^[0-9a-f]{40,64}$'
          readOnly: true
          description: Empty only when the worktree is missing or released.
        baseRef:
          type: string
          minLength: 1
          maxLength: 1024
          readOnly: true
        managed:
          type: boolean
          readOnly: true
          description: Whether the adapter created and owns cleanup of this worktree.
        clean:
          type: boolean
          readOnly: true
        state:
          $ref: '#/components/schemas/WorktreeState'
        sessionCount:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
        createdAt:
          type: string
          format: date-time
          pattern: 'Z$'
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          pattern: 'Z$'
          readOnly: true
...
    TerminalStatus:
      type: string
      description: starting and running are live; exited and failed are terminal.
      enum:
        - starting
        - running
        - exited
        - failed
    Terminal:
      type: object
      additionalProperties: false
      required:
        - id
        - worktreeId
        - status
        - shell
        - columns
        - rows
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        worktreeId:
          type: string
          format: uuid
          readOnly: true
        status:
          $ref: '#/components/schemas/TerminalStatus'
        shell:
          type: string
          minLength: 1
          maxLength: 4096
          readOnly: true
          description: Configured executable; clients cannot select it.
        columns:
          type: integer
          format: int32
          minimum: 20
          maximum: 400
        rows:
          type: integer
          format: int32
          minimum: 5
          maximum: 200
        exitCode:
          type: integer
          format: int32
          readOnly: true
          description: Present after a normal process exit.
        createdAt:
          type: string
          format: date-time
          pattern: 'Z$'
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          pattern: 'Z$'
          readOnly: true
    CreateTerminalRequest:
      type: object
      additionalProperties: false
      required:
        - columns
        - rows
      properties:
        columns:
          type: integer
          format: int32
          minimum: 20
          maximum: 400
        rows:
          type: integer
          format: int32
          minimum: 5
          maximum: 200
    ResizeTerminalRequest:
      type: object
      additionalProperties: false
      required:
        - columns
        - rows
      properties:
        columns:
          type: integer
          format: int32
          minimum: 20
          maximum: 400
        rows:
          type: integer
          format: int32
          minimum: 5
          maximum: 200
    TerminalInputRequest:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          type: string
          minLength: 1
          maxLength: 65536
          description: UTF-8 input written to the terminal's pseudo-terminal.
...
    TerminalList:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Terminal'
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
          description: Stable machine-readable reason, such as session_not_found.
        message:
          type: string
          minLength: 1
          maxLength: 2048
        details:
          type: object
          description: Optional per-code context, such as the offending field.
          additionalProperties: true
...
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The addressed resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The resource is not in a state that allows this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: The request could not be completed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'